+++ src/archetype.rs
@@ -115,12 +115,12 @@ impl Archetype {
     /// Get the `T` components of these entities, if present
     ///
     /// Useful for efficient serialization.
-    pub fn get<T: Component>(&self) -> Option<ColumnRef<'_, T>> {
+    pub fn get<T: Component>(&self) -> Option<ArchetypeColumn<'_, T>> {
         let state = self.get_state::<T>()?;
         let ptr = self.get_base::<T>(state);
         let column = unsafe { slice::from_raw_parts_mut(ptr.as_ptr(), self.len as usize) };
         self.borrow::<T>(state);
-        Some(ColumnRef {
+        Some(ArchetypeColumn {
             archetype: self,
             column,
         })
@@ -494,7 +494,11 @@ impl<V> OrderedTypeIdMap<V> {
     }
 }
 
-/// Metadata required to store a component
+/// Metadata required to store a component.
+///
+/// All told, this means a [`TypeId`], to be able to dynamically name/check the component type; a
+/// [`Layout`], so that we know how to allocate memory for this component type; and a drop function
+/// which internally calls [`core::ptr::drop_in_place`] with the correct type parameter.
 #[derive(Debug, Copy, Clone)]
 pub struct TypeInfo {
     pub id: TypeId,
@@ -505,7 +509,7 @@ pub struct TypeInfo {
 }
 
 impl TypeInfo {
-    /// Metadata for `T`
+    /// Construct a `TypeInfo` directly from the static type.
     pub fn of<T: 'static>() -> Self {
         unsafe fn drop_ptr<T>(x: *mut u8) {
             x.cast::<T>().drop_in_place()
@@ -520,16 +524,44 @@ impl TypeInfo {
         }
     }
 
-    pub(crate) fn id(&self) -> TypeId {
+    /// Construct a `TypeInfo` from its components. This is useful in the rare case that you have
+    /// some kind of pointer to raw bytes/erased memory holding a component type, coming from a
+    /// source unrelated to hecs, and you want to treat it as an insertable component by
+    /// implementing the `DynamicBundle` API.
+    pub fn from_parts(id: TypeId, layout: Layout, drop: unsafe fn(*mut u8)) -> Self {
+        Self {
+            id,
+            layout,
+            drop,
+            #[cfg(debug_assertions)]
+            type_name: "<unknown> (TypeInfo constructed from parts)",
+        }
+    }
+
+    /// Access the `TypeId` for this component type.
+    pub fn id(&self) -> TypeId {
         self.id
     }
 
-    pub(crate) fn layout(&self) -> Layout {
+    /// Access the `Layout` of this component type.
+    pub fn layout(&self) -> Layout {
         self.layout
     }
 
-    pub(crate) unsafe fn drop(&self, data: *mut u8) {
+    /// Directly call the destructor on a pointer to data of this component type.
+    ///
+    /// # Safety
+    ///
+    /// All of the caveats of [`core::ptr::drop_in_place`] apply, with the additional requirement
+    /// that this method is being called on a pointer to an object of the correct component type.
+    pub unsafe fn drop(&self, data: *mut u8) {
         (self.drop)(data)
+    }
+
+    /// Get the function pointer encoding the destructor for the component type this `TypeInfo`
+    /// represents.
+    pub fn drop_shim(&self) -> unsafe fn(*mut u8) {
+        self.drop
     }
 }
 
@@ -559,26 +591,26 @@ impl PartialEq for TypeInfo {
 impl Eq for TypeInfo {}
 
 /// Shared reference to a single column of component data in an [`Archetype`]
-pub struct ColumnRef<'a, T: Component> {
+pub struct ArchetypeColumn<'a, T: Component> {
     archetype: &'a Archetype,
     column: &'a [T],
 }
 
-impl<T: Component> Deref for ColumnRef<'_, T> {
+impl<T: Component> Deref for ArchetypeColumn<'_, T> {
     type Target = [T];
     fn deref(&self) -> &[T] {
         self.column
     }
 }
 
-impl<T: Component> Drop for ColumnRef<'_, T> {
+impl<T: Component> Drop for ArchetypeColumn<'_, T> {
     fn drop(&mut self) {
         let state = self.archetype.get_state::<T>().unwrap();
         self.archetype.release::<T>(state);
     }
 }
 
-impl<T: Component> Clone for ColumnRef<'_, T> {
+impl<T: Component> Clone for ArchetypeColumn<'_, T> {
     fn clone(&self) -> Self {
         let state = self.archetype.get_state::<T>().unwrap();
         self.archetype.borrow::<T>(state);
@@ -589,7 +621,7 @@ impl<T: Component> Clone for ColumnRef<'_, T> {
     }
 }
 
-impl<T: Component + fmt::Debug> fmt::Debug for ColumnRef<'_, T> {
+impl<T: Component + fmt::Debug> fmt::Debug for ArchetypeColumn<'_, T> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.column.fmt(f)
     }