+++ src/query.rs
@@ -569,8 +569,6 @@ impl<'w, Q: Query> QueryBorrow<'w, Q> {
     }
 
     /// Execute the query
-    ///
-    /// Must be called only once per query.
     // The lifetime narrowing here is required for soundness.
     pub fn iter(&mut self) -> QueryIter<'_, Q> {
         self.borrow();
@@ -925,7 +923,7 @@ impl<'q, Q: Query> Iterator for BatchedIter<'q, Q> {
     }
 }
 
-/// A sequence of entities yielded by `BatchedIter`
+/// A sequence of entities yielded by [`BatchedIter`]
 pub struct Batch<'q, Q: Query> {
     meta: &'q [EntityMeta],
     state: ChunkIter<Q>,